swiftship/services/customer-booking-service/requirements.txt
# Core FastAPI
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Caching
redis==5.0.4
aioredis==2.0.1

# Kafka
aiokafka==0.10.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9

# HTTP client (service-to-service calls)
httpx==0.27.0
aiohttp==3.9.5

# Object storage
minio==7.2.7

# Elasticsearch
elasticsearch[async]==8.13.0

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.3.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
anyio==4.4.0
